requirements.txt
asgiref==3.3.1
backcall
boto3==1.17.62
botocore==1.20.62
brotlipy==0.7.0
certifi==2020.12.5
cffi
chardet
colorama
cryptography
decorator
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna
ipykernel
ipython
ipython-genutils
jedi==0.17.0
jmespath==0.10.0
jupyter-client
jupyter-core
numpy==1.20.1
parso
pickleshare
Pillow==8.2.0
prompt-toolkit
psycopg2==2.8.6
pycparser
Pygments
pyOpenSSL
pypng==0.0.20
PyQRCode==1.2.1
PySocks
python-dateutil
pytz==2021.1
pyzmq==20.0.0
requests
s3transfer==0.4.2
six
sqlparse==0.4.1
timezone==0.0.1
tornado
traitlets
urllib3
wcwidth
whitenoise==5.2.0
win-inet-pton
wincertstore==0.2
